Attendance App/requirements.txt
pandas==2.1.4
Pillow==10.1.0
ttkthemes==3.2.2
openpyxl==3.1.2
pyinstaller==6.3.0 